requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-mysql==4.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.3
requests==2.27.1
Shapely==1.8.2
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.9
